backend/pyproject.toml
[project]
name = "chat-ui"
version = "0.1.0"
description = "A simple AI agent with UI rendered responses."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.123.7",
    "langchain[openai]>=1.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "yfinance>=0.2.66",
]
